requirements.txt
requests
tdqm
python-nmap
